f4e6

+++ fase04/ex06/passadoOuFuturo.py
@@ -1,5 +1,5 @@
 def passadoOuFuturo(tempo):
-    if int(tempo/1000) < 1445385600:
+    if tempo/1000 < 1445385600:
         return "Passado"
     else:
         return "Futuro"
